refactor: extract non-lister functions to services
